refactor ChatMessageToolbar
